Maintenance end node to SinglyLinkedList
For performance concern when the list becomes long.
Add empty() function
when add new node before the head node, append it to the end of the list or add it in the middle of the list, maintenance the fields of head node and end node
and the size of list.

Fix some bugs and add releated junit test.